refactor: move back button and BookAdditionChoiceScreen
